some changes to when term is calculated
